move definitions to their own folder
