+++ definitions/index.ts
@@ -1,3 +1,3 @@
 // ANCHOR: Export User Definition
-export { userLoginDefinition } from './modules/userLoginDefinition';
-export { userRegisterDefinition } from './modules/userRegisterDefinition';
+export { userLoginDefinition } from './modules/user/userLoginDefinition';
+export { userRegisterDefinition } from './modules/user/userRegisterDefinition';
