+++ definitions/modules/user/userRegisterDefinition.ts
